requirements.txt
fastapi
sqlalchemy
uvicorn
bcrypt
python-jose[cryptography]
python-multipart